Изменен файл генерации данных. Создан репозиторий фильмов, к которому обращаются фрагменты
