+++ app/src/main/java/com/alexanderpodkopaev/androidacademyproject/FragmentMoviesList.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.alexanderpodkopaev.androidacademyproject.UiUtils.calculateNoOfColumns
 import com.alexanderpodkopaev.androidacademyproject.adapter.MovieClickListener
 import com.alexanderpodkopaev.androidacademyproject.adapter.MoviesAdapter
-import com.alexanderpodkopaev.androidacademyproject.data.getMovies
+import com.alexanderpodkopaev.androidacademyproject.data.MoviesRepository
 
 class FragmentMoviesList : Fragment(), MovieClickListener {
 
@@ -22,7 +22,7 @@ class FragmentMoviesList : Fragment(), MovieClickListener {
     ): View? {
         val view = inflater.inflate(R.layout.fragment_movies_list, container, false)
         val adapter = MoviesAdapter()
-        adapter.bindMovies(generatedMovies)
+        adapter.bindMovies(MoviesRepository().getMovies())
         adapter.onMovieClickListener = this
         val recyclerViewMovies = view.findViewById<RecyclerView>(R.id.rvMoviesList)
         recyclerViewMovies?.layoutManager =
@@ -45,10 +45,6 @@ class FragmentMoviesList : Fragment(), MovieClickListener {
             ?.replace(R.id.flFragment, FragmentMoviesDetails.newInstance(movieTitle))
             ?.addToBackStack(null)
             ?.commit()
-    }
-
-    companion object {
-        var generatedMovies = getMovies().plus(getMovies())
     }
 
 }
